[vgui_controls] Use safe str* APIs

+++ public/vgui_controls/consoledialog.h
@@ -58,7 +58,7 @@ public:
 	void ColorPrint( const Color& clr, const char *pMessage ) override;
 	void Print( const char *pMessage ) override;
 	void DPrint( const char *pMessage ) override;
-	virtual void GetConsoleText( char *pchText, size_t bufSize ) const;
+	virtual void GetConsoleText( char *pchText, intp bufSize ) const;
 
 	// clears the console
 	void Clear();
